requirements.txt
python-dotenv==1.0.0
requests==2.31.0
openai-agents==0.0.3
rich==13.7.0 